+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/indicators_flyout/indicators_flyout.test.tsx
@@ -15,7 +15,7 @@ import { mockUiSetting } from '../../../../common/mocks/mock_kibana_ui_settings_
 import { TestProvidersComponent } from '../../../../common/mocks/test_providers';
 import { generateFieldTypeMap } from '../../../../common/mocks/mock_field_type_map';
 import { unwrapValue } from '../../lib/unwrap_value';
-import { getDisplayName } from '../../lib/display_value';
+import { getDisplayName } from '../../lib/display_name';
 
 const mockIndicator = generateMockIndicator();
 const mockFieldTypesMap = generateFieldTypeMap();
