+++ x-pack/plugins/threat_intelligence/public/modules/timeline/components/add_to_timeline/add_to_timeline.tsx
@@ -10,7 +10,7 @@ import { DataProvider, QueryOperator } from '@kbn/timelines-plugin/common';
 import { AddToTimelineButtonProps } from '@kbn/timelines-plugin/public';
 import { EuiButtonEmpty, EuiButtonIcon } from '@elastic/eui/src/components/button';
 import { EMPTY_VALUE } from '../../../../../common/constants';
-import { getDisplayName } from '../../../indicators/lib/display_value';
+import { getDisplayName } from '../../../indicators/lib/display_name';
 import { ComputedIndicatorFieldId } from '../../../indicators/components/indicators_table/cell_renderer';
 import { useKibana } from '../../../../hooks/use_kibana';
 import { unwrapValue } from '../../../indicators/lib/unwrap_value';
